Rename button component
RISDEV-136
